Improvements

+++ Newsier.Application/ViewModels/PublisherVm.cs
@@ -14,15 +14,15 @@ namespace Newsier.Application.ViewModels
         public string Email { get; set; }
         public string Role { get; set; }
         public string CreatedAt { get; set; }
-        public uint Publications { get; set; }
-        public uint Comments { get; set; }
+        public int Publications { get; set; }
+        public int Comments { get; set; }
 
         public void Mapping(Profile profile)
         {
             profile.CreateMap<Publisher, PublisherVm>()
                 .ForMember(x => x.CreatedAt, opt => opt.MapFrom(x => $"{x.CreatedAt:MMMM dd}{(x.CreatedAt.Year == DateTime.UtcNow.Year ? "" : $"{x.CreatedAt:, yyyy}")}"))
-                .ForMember(x => x.Publications, opt => opt.MapFrom(x => (uint)x.Publications.Count))
-                .ForMember(x => x.Comments, opt => opt.MapFrom(x => (uint)x.Comments.Count));
+                .ForMember(x => x.Publications, opt => opt.MapFrom(x => x.Publications.Count))
+                .ForMember(x => x.Comments, opt => opt.MapFrom(x => x.Comments.Count));
         }
     }
 }
